+++ changelog.md
@@ -66,3 +66,4 @@ routes: cart
 ###[7.0.0]
 -added appstate,store, product states, reducer, action
 - added selectors,effects, entites 
+- added selector By prodId and dtore go in product-resolve guard and in product-form component